Fix formatting (separate songcontext and provider)
