--- frontend/lyric-sync-app/src/contexts/SongContextHooks.ts
@@ -1,10 +1,0 @@
-import { useContext } from 'react';
-import { SongContext } from './SongContext';
-
-export const useSong = () => {
-  const context = useContext(SongContext);
-  if (!context) {
-    throw new Error('useSong must be used within a SongProvider');
-  }
-  return context;
-};
